# [GraphBolt][CUDA] Async sample neighbors and compaction.