--- a/python/dgl/graphbolt/impl/fused_csc_sampling_graph.py
+++ b/python/dgl/graphbolt/impl/fused_csc_sampling_graph.py
@@ -21,6 +21,22 @@ __all__ = [
 ]
 
 
+class _SampleNeighborsWaiter:
+    def __init__(self, fn, future, seed_offsets):
+        self.fn = fn
+        self.future = future
+        self.seed_offsets = seed_offsets
+
+    def wait(self):
+        """Returns the stored value when invoked."""
+        fn = self.fn
+        C_sampled_subgraph = self.future.wait()
+        seed_offsets = self.seed_offsets
+        # Ensure there is no memory leak.
+        self.fn = self.future = self.seed_offsets = None
+        return fn(C_sampled_subgraph, seed_offsets)
+
+
 class FusedCSCSamplingGraph(SamplingGraph):
     r"""A sampling graph in CSC format."""
 
@@ -696,6 +712,7 @@ class FusedCSCSamplingGraph(SamplingGraph):
         replace: bool = False,
         probs_name: Optional[str] = None,
         returning_indices_is_optional: bool = False,
+        is_asynchronous: bool = False,
     ) -> SampledSubgraphImpl:
         """Sample neighboring edges of the given nodes and return the induced
         subgraph.
@@ -739,6 +756,9 @@ class FusedCSCSamplingGraph(SamplingGraph):
             Boolean indicating whether it is okay for the call to this function
             to leave the indices tensor uninitialized. In this case, it is the
             user's responsibility to gather it using the edge ids.
+        is_asynchronous: bool
+            Boolean indicating whether the call is asynchronous. If so, the
+            result can be obtained by calling wait on the returned future.
 
         Returns
         -------
@@ -784,9 +804,16 @@ class FusedCSCSamplingGraph(SamplingGraph):
             probs_or_mask=probs_or_mask,
             returning_indices_is_optional=returning_indices_is_optional,
         )
-        return self._convert_to_sampled_subgraph(
-            C_sampled_subgraph, seed_offsets
-        )
+        if is_asynchronous:
+            return _SampleNeighborsWaiter(
+                self._convert_to_sampled_subgraph,
+                C_sampled_subgraph,
+                seed_offsets,
+            )
+        else:
+            return self._convert_to_sampled_subgraph(
+                C_sampled_subgraph, seed_offsets
+            )
 
     def _check_sampler_arguments(self, nodes, fanouts, probs_or_mask):
         if nodes is not None:
@@ -835,6 +862,7 @@ class FusedCSCSamplingGraph(SamplingGraph):
         replace: bool = False,
         probs_or_mask: Optional[torch.Tensor] = None,
         returning_indices_is_optional: bool = False,
+        is_asynchronous: bool = False,
     ) -> torch.ScriptObject:
         """Sample neighboring edges of the given nodes and return the induced
         subgraph.
@@ -877,6 +905,9 @@ class FusedCSCSamplingGraph(SamplingGraph):
             Boolean indicating whether it is okay for the call to this function
             to leave the indices tensor uninitialized. In this case, it is the
             user's responsibility to gather it using the edge ids.
+        is_asynchronous: bool
+            Boolean indicating whether the call is asynchronous. If so, the
+            result can be obtained by calling wait on the returned future.
 
         Returns
         -------
@@ -885,7 +916,12 @@ class FusedCSCSamplingGraph(SamplingGraph):
         """
         # Ensure nodes is 1-D tensor.
         self._check_sampler_arguments(seeds, fanouts, probs_or_mask)
-        return self._c_csc_graph.sample_neighbors(
+        sampling_fn = (
+            self._c_csc_graph.sample_neighbors_async
+            if is_asynchronous
+            else self._c_csc_graph.sample_neighbors
+        )
+        return sampling_fn(
             seeds,
             seed_offsets,
             fanouts.tolist(),
@@ -906,6 +942,7 @@ class FusedCSCSamplingGraph(SamplingGraph):
         returning_indices_is_optional: bool = False,
         random_seed: torch.Tensor = None,
         seed2_contribution: float = 0.0,
+        is_asynchronous: bool = False,
     ) -> SampledSubgraphImpl:
         """Sample neighboring edges of the given nodes and return the induced
         subgraph via layer-neighbor sampling from the NeurIPS 2023 paper
@@ -1024,7 +1061,12 @@ class FusedCSCSamplingGraph(SamplingGraph):
             seed_offsets = self._indptr_node_type_offset_list
         probs_or_mask = self.edge_attributes[probs_name] if probs_name else None
         self._check_sampler_arguments(seeds, fanouts, probs_or_mask)
-        C_sampled_subgraph = self._c_csc_graph.sample_neighbors(
+        sampling_fn = (
+            self._c_csc_graph.sample_neighbors_async
+            if is_asynchronous
+            else self._c_csc_graph.sample_neighbors
+        )
+        C_sampled_subgraph = sampling_fn(
             seeds,
             seed_offsets,
             fanouts.tolist(),
@@ -1035,9 +1077,16 @@ class FusedCSCSamplingGraph(SamplingGraph):
             random_seed,
             seed2_contribution,
         )
-        return self._convert_to_sampled_subgraph(
-            C_sampled_subgraph, seed_offsets
-        )
+        if is_asynchronous:
+            return _SampleNeighborsWaiter(
+                self._convert_to_sampled_subgraph,
+                C_sampled_subgraph,
+                seed_offsets,
+            )
+        else:
+            return self._convert_to_sampled_subgraph(
+                C_sampled_subgraph, seed_offsets
+            )
 
     def temporal_sample_neighbors(
         self,